Moved functions from random.py into random plugin. Removed random.py
